모델 타입 통일

+++ models/Subject.js
@@ -2,7 +2,7 @@ const { DataTypes } = require("sequelize");
 const sequelize = require("../config/database");
 
 const Subject = sequelize.define("Subject", {
-  id: { type: DataTypes.BIGINT, autoIncrement: true, primaryKey: true },
+  id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
   name: { type: DataTypes.STRING(50), allowNull: false },     // 국어, 수학, 영어, 탐구
   category: { type: DataTypes.STRING(50), allowNull: false }, // 공통 / 선택
 }, {
